requirements.txt
Unidecode == 1.1.2
discord.py == 1.6.0
discord_py_slash_command == 1.1.2
